# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Parallel AI Coding Orchestrator API",
    "description": "\n## Week 2 MVP - Manager AI Integration API\n\nEnterprise-grade REST API for orchestrating and monitoring parallel AI coding workflows.\n\n### Key Features\n- **Supervisor API**: Monitor and control Claude Code worker instances\n- **Resource Management**: Hierarchical resource allocation with depth-based quotas\n- **Job Orchestration**: Manage job lifecycle with state machine validation\n- **Authentication**: JWT-based authentication with scope-based authorization\n\n### Architecture\n- **Database**: SQLite (development) / PostgreSQL (production)\n- **Authentication**: JWT Bearer tokens with scopes\n- **State Management**: Validated state machines for workers and jobs\n- **Testing**: 44 integration tests with 29% code coverage\n\n### Quick Start\n1. Generate test token: Use `/api/auth/token` endpoint\n2. Access protected endpoints with `Authorization: Bearer <token>` header\n3. Explore interactive docs at `/docs` or `/redoc`\n\n### Support\n- **GitHub**: [parallel-coding](https://github.com/your-org/parallel-coding)\n- **Documentation**: [Wiki](https://github.com/your-org/parallel-coding/wiki)\n",
    "contact": {
      "name": "Parallel AI Development Team",
      "email": "dev@parallel-ai.example.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "2.0.0-week2-mvp"
  },
  "paths": {
...
    "/api/resources/quotas": {
      "get": {
        "tags": [
          "resources"
        ],
        "summary": "Get resource quotas",
        "description": "Retrieve resource quotas configured for each hierarchy depth level.\n\n    **Hierarchical Resource Model:**\n    - Depth 0 (root): Maximum workers available\n    - Depth 1-5: Decreasing quotas for nested jobs\n\n    **Default Configuration:**\n    ```\n    Depth 0: 10 workers\n    Depth 1: 8 workers\n    Depth 2: 5 workers\n    Depth 3: 3 workers\n    Depth 4: 2 workers\n    Depth 5: 1 worker\n    ```\n\n    **Required Scope:** `resources:read`\n\n    **Example Response:**\n    ```json\n    {\n      \"quotas\": [\n        {\"depth\": 0, \"max_workers\": 10},\n        {\"depth\": 1, \"max_workers\": 8},\n        {\"depth\": 2, \"max_workers\": 5}\n      ]\n    }\n    ```",
        "operationId": "get_quotas_api_resources_quotas_get",
        "responses": {
          "200": {
            "description": "Resource quotas by depth level",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotasResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/resources/allocate": {
      "post": {
        "tags": [
          "resources"
        ],
        "summary": "Allocate Resources",
        "description": "Allocate resources for a job at a specific depth.",
        "operationId": "allocate_resources_api_resources_allocate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AllocateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllocationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/resources/release": {
      "post": {
        "tags": [
          "resources"
        ],
        "summary": "Release Resources",
        "description": "Release any resources held by a job at a depth.",
        "operationId": "release_resources_api_resources_release_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/resources/usage": {
      "get": {
        "tags": [
          "resources"
        ],
        "summary": "Get Usage",
        "description": "Get current resource usage by depth (allocated and available).",
        "operationId": "get_usage_api_resources_usage_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AllocateRequest": {
        "properties": {
          "job_id": {
            "type": "string",
            "minLength": 1,
            "title": "Job Id"
          },
          "depth": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Depth"
          },
          "worker_count": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Worker Count",
            "description": "Requested number of workers"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "depth",
          "worker_count"
        ],
        "title": "AllocateRequest"
      },
      "AllocationResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "depth": {
            "type": "integer",
            "title": "Depth"
          },
          "requested": {
            "type": "integer",
            "title": "Requested"
          },
          "granted": {
            "type": "integer",
            "title": "Granted"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "depth",
          "requested",
          "granted"
        ],
        "title": "AllocationResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuotaItem": {
        "properties": {
          "depth": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Depth"
          },
          "max_workers": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Max Workers"
          }
        },
        "type": "object",
        "required": [
          "depth",
          "max_workers"
        ],
        "title": "QuotaItem"
      },
      "QuotasResponse": {
        "properties": {
          "quotas": {
            "items": {
              "$ref": "#/components/schemas/QuotaItem"
            },
            "type": "array",
            "title": "Quotas"
          }
        },
        "type": "object",
        "required": [
          "quotas"
        ],
        "title": "QuotasResponse"
      },
      "ReleaseRequest": {
        "properties": {
          "job_id": {
            "type": "string",
            "minLength": 1,
            "title": "Job Id"
          },
          "depth": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Depth"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "depth"
        ],
        "title": "ReleaseRequest"
      },
      "ReleaseResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "depth": {
            "type": "integer",
            "title": "Depth"
          },
          "released": {
            "type": "boolean",
            "title": "Released"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "depth",
          "released"
        ],
        "title": "ReleaseResponse"
      },
...
      "UsageItem": {
        "properties": {
          "depth": {
            "type": "integer",
            "title": "Depth"
          },
          "allocated": {
            "type": "integer",
            "title": "Allocated"
          },
          "available": {
            "type": "integer",
            "title": "Available"
          }
        },
        "type": "object",
        "required": [
          "depth",
          "allocated",
          "available"
        ],
        "title": "UsageItem"
      },
      "UsageResponse": {
        "properties": {
          "usage": {
            "items": {
              "$ref": "#/components/schemas/UsageItem"
            },
            "type": "array",
            "title": "Usage"
          }
        },
        "type": "object",
        "required": [
          "usage"
        ],
        "title": "UsageResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}